♻️ Refactor_Extract foodStockButton widget
